Add helm chart

+++ src/main/java/net/checkconsulting/scpiinvestapi/dto/ScpiBatchDto.java
@@ -13,10 +13,10 @@ public class ScpiBatchDto {
 
     private String name;
     private Integer minimumSubscription;
-    private Integer capitalization;
+    private Long capitalization;
     private String manager;
-    private Integer subscriptionFees;
-    private Integer managementFees;
+    private Float subscriptionFees;
+    private Float managementFees;
     private Integer delayBenefit;
     private String rentFrequency;
     private String distributionRate;
