+++ src/main/java/net/checkconsulting/scpiinvestapi/entity/Scpi.java
@@ -23,10 +23,10 @@ public class Scpi {
     private Integer id;
     private String name;
     private Integer minimumSubscription;
-    private Integer capitalization;
+    private Long capitalization;
     private String manager;
-    private Integer subscriptionFees;
-    private Integer managementFees;
+    private Float subscriptionFees;
+    private Float managementFees;
     private Integer delayBenefit;
     private String rentFrequency;
